chore: refactor code so that it doesn't clutter app.js
