+++ src/app.js
@@ -1,51 +1,16 @@
 import Fastify from 'fastify';
-import { getTodos, saveTodos } from './utils/db/index.js';
-import { v4 } from 'uuid';
+
+import { createTodo } from './services/todos/create-todo.js';
+import { general } from './services/general/index.js';
 
 const prefix = '/api';
 
 export async function build () {
   const fastify = Fastify({ logger: true });
 
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+  fastify.get(prefix, general);
 
-  fastify.post(`${prefix}/todo`, async (request, reply) => {
-    const { body } = request;
-    const { title, description, isDone = false } = body;
-    const db = await getTodos();
-
-    const id = v4();
-
-    const todo = {
-      title,
-      description,
-      isDone,
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.todos[id] = todo;
-
-    await saveTodos(db);
-
-    /**
-     * const newObj = {
-     *   id
-     * }
-     *
-     * for (const key in todo) {
-     *   newObj[key] = todo[key]
-     * }
-     *
-     * return newObj
-     */
-    return {
-      id,
-      ...todo
-    };
-  });
+  fastify.post(`${prefix}/todo`, createTodo);
 
   return fastify;
 }
