+++ src/services/general/index.js
@@ -1,0 +1,4 @@
+
+export const general = async (request, reply) => {
+  return { success: true };
+};
